ModalForm and common modal components were added. Post and dashboard components were refactored for work with new modal.
