+++ src/components/common/textarea/Textarea.scss
@@ -3,7 +3,7 @@
   padding: 10px;
   font-weight: 400;
   font-size: 16px;
-  line-height: 1;
+  //line-height: 1;
   color: #495057;
   background-color: #fff;
   background-clip: padding-box;
